[update] soundJson修正
