--- GFF2026A/Assets/Scripts/UI/BGMSEUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: de3e6c25434566c4c9912cde8232fed1
