--- GFF2026A/Assets/Scripts/UI/OptionScreen.cs
@@ -1,24 +1,0 @@
-using Unity.VisualScripting;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class OptionScreen : MonoBehaviour
-{
-    // 閉じるボタン
-    [SerializeField] private Button closeButton;
-
-    // オプション画面のオブジェクト
-    [SerializeField] private GameObject optionScreen;
-
-    // 閉じるボタンにクリックイベントを追加
-    void Start()
-    {
-        closeButton.onClick.AddListener(ClosePopup);
-    }
-
-    // ポップアップを閉じる
-    void ClosePopup()
-    {
-        Destroy(optionScreen);
-    }
-}
