--- GFF2026A/Assets/Scripts/UI/OptionUI.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cff35492e83e1674ea9e7e3f84501aed
